+++ lib/gundam.rb
@@ -2,6 +2,7 @@ require 'English' # needed by $CHILD_STATUS
 require 'forwardable'
 
 require_relative 'gundam/configurable'
+require_relative 'gundam/helpers/colorize_helper'
 require_relative 'gundam/commands/command'
 require_relative 'gundam/commands/plugin'
 require_relative 'gundam/commands/shared/file_helper'
