+++ lib/gundam/decorators/decorator.rb
@@ -1,5 +1,3 @@
-require_relative '../../helpers/colorize_helper'
-
 class Decorator < SimpleDelegator
-  include ColorizeHelper
+  include Gundam::ColorizeHelper
 end
