+++ lib/gundam/error_handler.rb
@@ -1,5 +1,3 @@
-require_relative '../helpers/colorize_helper'
-
 module Gundam
   class ErrorHandler
     class << self
