+++ spec/lib/gundam/context/with_repository_spec.rb
@@ -71,9 +71,9 @@ describe Gundam::Context::WithRepository do
 
       it 'returns the service' do
         service = double('Gundam::Github::API::V3::Gateway')
-        factory = double('Gundam::PlatformServiceFactory')
+        factory = double('Gundam::RepoServiceFactory')
 
-        allow(PlatformServiceFactory).to receive(:with_platform).with('Github')
+        allow(Gundam::RepoServiceFactory).to receive(:with_platform).with('Github')
           .and_return(factory)
         allow(factory).to receive(:build).and_return(service)
